components/apps/guide/src/requirements.txt
grpcio==1.20.1
grpcio-tools==1.20.1
numpy==1.16.3
protobuf==3.7.1
six==1.12.0
